feat: refactor timeline components to use custom hooks for state management and event handling
