+++ src/components/TimelineLane.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { useTimelineItemUpdate } from '../hooks/useTimelineItemUpdate';
 import { TimelineConfig, TimelineItem } from '../types/timeline';
 import TimelineItemComponent from './TimelineItemComponent';
 
@@ -13,6 +14,8 @@ const TimelineLane: React.FC<TimelineLaneProps> = ({
   items,
   config
 }) => {
+  const handleUpdateDates = useTimelineItemUpdate();
+
   return (
     <div
       className="relative border-b border-gray-200"
@@ -27,13 +30,7 @@ const TimelineLane: React.FC<TimelineLaneProps> = ({
           key={item.id}
           item={item}
           config={config}
-          onUpdateDates={(id, newStart, newEnd) => {
-            if (typeof window !== 'undefined' && window.dispatchEvent) {
-              window.dispatchEvent(new CustomEvent('timelineItemUpdate', {
-                detail: { id, newStart, newEnd }
-              }));
-            }
-          }}
+          onUpdateDates={handleUpdateDates}
         />
       ))}
     </div>
