+++ src/hooks/useHover.ts
@@ -1,0 +1,14 @@
+import { useState, useCallback } from 'react';
+
+export function useHover() {
+  const [isHovered, setIsHovered] = useState(false);
+
+  const handleMouseEnter = useCallback(() => setIsHovered(true), []);
+  const handleMouseLeave = useCallback(() => setIsHovered(false), []);
+
+  return {
+    isHovered,
+    handleMouseEnter,
+    handleMouseLeave
+  };
+}
